Start tests for mongo, resolvers, move resolvers to own file
