+++ .gitignore
@@ -47,3 +47,6 @@ certs/
 saml.pem
 
 tsconfig.tsbuildinfo
+
+# generated by jest mongodb
+globalConfig.json
