+++ jest-mongodb-config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  mongodbMemoryServerOptions: {
+    binary: {
+      version: '4.0.0',
+      skipMD5: true,
+    },
+    autostart: false,
+    instance: {},
+  },
+}
