+++ jest.config.js
@@ -6,6 +6,7 @@ module.exports = {
     '^.+\\.(js|jsx|ts|tsx)$': 'babel-jest',
   },
   transformIgnorePatterns: ['node_modules/(?!(.keystone)/)'],
+  watchPathIgnorePatterns: ['globalconfig'],
   moduleNameMapper: {
     '\\.(scss|sass|css)$': 'identity-obj-proxy',
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
@@ -21,6 +22,7 @@ module.exports = {
     '^types/(.*)': '<rootDir>/src/types/$1',
   },
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+  preset: '@shelf/jest-mongodb',
   coverageThreshold: {
     global: {
       statements: 95,
