+++ package.json
@@ -60,6 +60,7 @@
   },
   "devDependencies": {
     "@babel/core": "7.16.0",
+    "@shelf/jest-mongodb": "^2.1.0",
     "@storybook/addon-a11y": "6.3.12",
     "@storybook/addon-actions": "6.3.12",
     "@storybook/addon-essentials": "6.3.12",
